refactor: Consolidate duplicate ensure_initial_message_exists
Move the single canonical version into chat_messages.utils and delete
both copies from users.utils and coach.services. All callers and test
mock paths updated to the new location.

Made-with: Cursor
