+++ server/apps/coach/services/coach_service/utils/__init__.py
@@ -2,12 +2,10 @@ from .apply_coach_response_actions import apply_coach_response_actions
 from .apply_user_component_actions import apply_user_component_actions
 from .build_coach_prompt import build_coach_prompt
 from .build_coach_response_data import build_coach_response_data
-from .ensure_initial_message_exists import ensure_initial_message_exists
 from .generate_coach_ai_response import generate_coach_ai_response
 from .get_recent_chat_messages_for_prompt import get_recent_chat_messages_for_prompt
 
 __all__ = [
-    "ensure_initial_message_exists",
     "apply_user_component_actions",
     "build_coach_prompt",
     "get_recent_chat_messages_for_prompt",
