+++ server/apps/users/tests/test_get_user_chat_messages.py
@@ -49,31 +49,30 @@ class GetUserChatMessagesFunctionTests(TestCase):
         self.assertEqual(messages[1].content, "Second message")
         self.assertEqual(messages[2].content, "Third message")
 
-    @patch("apps.users.utils.ensure_initial_message_exists.get_initial_message")
-    def test_adds_initial_message_if_empty(self, mock_get_initial):
+    @patch(
+        "apps.chat_messages.utils.ensure_initial_message_exists.INITIAL_MESSAGE",
+        "Welcome to coaching!",
+    )
+    def test_adds_initial_message_if_empty(self):
         """Test that initial message is added if chat is empty."""
-        mock_get_initial.return_value = "Welcome to coaching!"
-
         messages = get_user_chat_messages(self.user)
 
         self.assertEqual(len(messages), 1)
         self.assertEqual(messages[0].content, "Welcome to coaching!")
         self.assertEqual(messages[0].role, MessageRole.COACH)
 
-    @patch("apps.users.utils.ensure_initial_message_exists.get_initial_message")
-    def test_no_initial_message_if_none_configured(self, mock_get_initial):
+    @patch(
+        "apps.chat_messages.utils.ensure_initial_message_exists.INITIAL_MESSAGE",
+        "",
+    )
+    def test_no_initial_message_if_none_configured(self):
         """Test behavior when no initial message is configured."""
-        mock_get_initial.return_value = None
-
         messages = get_user_chat_messages(self.user)
 
         self.assertEqual(len(messages), 0)
 
-    @patch("apps.users.utils.ensure_initial_message_exists.get_initial_message")
-    def test_does_not_add_initial_if_messages_exist(self, mock_get_initial):
+    def test_does_not_add_initial_if_messages_exist(self):
         """Test that initial message is not added if user already has messages."""
-        mock_get_initial.return_value = "Welcome!"
-
         ChatMessage.objects.create(
             user=self.user,
             content="Existing message",
@@ -84,8 +83,6 @@ class GetUserChatMessagesFunctionTests(TestCase):
 
         self.assertEqual(len(messages), 1)
         self.assertEqual(messages[0].content, "Existing message")
-        # get_initial_message should not be called if messages exist
-        mock_get_initial.assert_not_called()
 
     def test_does_not_return_other_users_messages(self):
         """Test that function only returns messages for the specified user."""
