+++ server/apps/users/utils/__init__.py
@@ -3,7 +3,3 @@ Users app utilities.
 
 This package contains utility functions shared across the users app.
 """
-
-from apps.users.utils.ensure_initial_message_exists import ensure_initial_message_exists
-
-__all__ = ["ensure_initial_message_exists"]
